test(search): relocate 7 command/CLI search-cluster suites to integration (#664 C3.2)
These 7 files all drive search/curate/show through the COMMAND layer
(akmSearch / akmCurate / searchForCuration / akmShowUnified / the full CLI via
runCliCapture), each building a DISTINCT on-disk corpus per test. None call
searchLocal/searchOnDb directly, so the new :memory: searchOnDb({db}) seam
cannot purify them without discarding the source-resolution / registry-merge /
CLI-arg / config coverage that IS their subject — and a single beforeAll index
won't serve them because each test seeds different stash content. Per the #664
plan ("if a file genuinely cannot avoid a per-test rebuild and isn't a clean
beforeAll candidate, RELOCATE to tests/integration/"), all 7 move to the
integration tier where their per-test full index rebuild is correct.

show-indexer-parity's subject is indexer.lookup ↔ akmShowUnified ↔ on-disk path
parity (not search ranking); it is integration by nature.

Bookkeeping:
- Removed all 7 from ALLOWED_FULL_INDEX; UNIT_PURITY_BASELINE 39 → 32.
- Updated the 4 ENV_ASSIGN_ALLOWED entries in lint-tests-isolation.ts to their
  new tests/integration/ paths (the isolation lint scans integration too).
- Fixed relative import depth (+1 ../) for the 5 files moved from tests/, plus
  the __dirname fixtures path in curate-relevance-eval.

All 53 relocated tests pass in their new location; lint + ratchet meta-tests green.
