+++ scripts/lint-tests-unit-purity.ts
@@ -90,16 +90,9 @@ const ALLOWED_FULL_INDEX = new Set<string>([
   "tests/commands/improve/improve-dry-run-side-effects.test.ts",
   "tests/commands/improve/improve-sync.test.ts",
   "tests/commands/scope-flags.test.ts",
-  "tests/commands/search.test.ts",
-  "tests/commands/show-indexer-parity.test.ts",
-  "tests/curate-relevance-eval.test.ts",
-  "tests/curate-search-for-curation.test.ts",
   "tests/env.test.ts",
   "tests/index-clean.test.ts",
   "tests/issue-36-repro.test.ts",
-  "tests/scoring-pipeline.test.ts",
-  "tests/search-include-proposed-cli.test.ts",
-  "tests/search-source-filter.test.ts",
   "tests/secret-indexing.test.ts",
   "tests/semantic-search-e2e.test.ts",
   "tests/session-indexing.test.ts",
@@ -113,7 +106,7 @@ const ALLOWED_FULL_INDEX = new Set<string>([
  * change whenever you remove an entry; never raise it. Meta-test:
  * `tests/lint-unit-purity-ratchet.test.ts`.
  */
-export const UNIT_PURITY_BASELINE = 39;
+export const UNIT_PURITY_BASELINE = 32;
 
 export function combinedAllowlistSize(): number {
   return ALLOWED_SERVE.size + ALLOWED_SPAWN.size + ALLOWED_FULL_INDEX.size;
