+++ tests/integration/search-include-proposed-cli.test.ts
@@ -17,9 +17,9 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 
-import { resetConfigCache, saveConfig } from "../src/core/config/config";
-import { akmIndex } from "../src/indexer/indexer";
-import { runCliCapture } from "./_helpers/cli";
+import { resetConfigCache, saveConfig } from "../../src/core/config/config";
+import { akmIndex } from "../../src/indexer/indexer";
+import { runCliCapture } from "../_helpers/cli";
 import {
   type Cleanup,
   sandboxStashDir,
@@ -27,7 +27,7 @@ import {
   sandboxXdgConfigHome,
   sandboxXdgDataHome,
   withEnv,
-} from "./_helpers/sandbox";
+} from "../_helpers/sandbox";
 
 const tempDirs: string[] = [];
