+++ tests/integration/show-indexer-parity.test.ts
